Use relaxed error handler for json deserializer
The error handler ignores deserialization errors on nullable properties.
We can mark all properties which are not strictly neccessary for the
functionality of this plugin as nullable, and improve the user
experience by having less failures in loading content.

We might have to refactor the xtream client to include logging, as
errors are now silently ignored.